backend/requirements.txt
# 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# 数据库
sqlalchemy==2.0.23
aiosqlite==0.19.0

# 数据验证
pydantic==2.5.2
pydantic-settings==2.1.0

# 安全认证
bcrypt==4.1.2
PyJWT==2.8.0

# 2FA 两步验证
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow==10.1.0

# 测试
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2

# 行情数据
tushare==1.4.25
pandas==2.0.3

# 工具
email-validator==2.1.0
